requirements_production.txt
# TITAN ERP - Production Requirements
# Ever Built Production Deployment

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12

# Database & BigQuery
google-cloud-bigquery==3.25.0
google-auth==2.35.0

# AI & ML
google-generativeai==0.8.3
Pillow==10.4.0

# Data Processing
pandas>=2.2.0,<3.0.0  # Lock to <3.0 for BigQuery/Streamlit compatibility
numpy>=1.26.0  # Python 3.13 compatible

# Scheduling (Auto-Pilot)
APScheduler==3.10.4
tzlocal==5.2

# Excel & Drive Integration
openpyxl==3.1.5
gspread==6.1.2
google-api-python-client==2.147.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3

# Streamlit (Admin Dashboard)
streamlit==1.39.0
plotly==5.24.1

# Development
pytest==8.3.3
black==24.10.0
